aider: Change JSON output to be indented with 4 spaces for each level of nesting.
# Aider chat conversation:

ASSISTANT: To achieve this, we need to use the `json.MarshalIndent` function instead of `json.Marshal`. This function takes an additional parameter which specifies the prefix for each line of output and another parameter for the indentation of each element.

Here is the change that needs to be made in the `cmd/root.go` file:

```go
cmd/root.go
<<<<<<< HEAD
			jsonData, err := json.Marshal(LogEntries)
			if err != nil {
				log.Fatal(err)
			}
			fmt.Println(string(jsonData))
=======
			jsonData, err := json.MarshalIndent(LogEntries, "", "    ")
			if err != nil {
				log.Fatal(err)
			}
			fmt.Println(string(jsonData))
>>>>>>> updated
```

This change will make the JSON output indented with 4 spaces for each level of nesting.