aider: Implement the 'set' subcommand in cmd/set.go to handle flag values, toggling boolean flags, and printing flag states.
# Aider chat conversation:

USER: please implement the 'set' sub command in the cmd/set.go file
this command will deal with the values of the persistant flags  in rootCmd.
If we supply two arguments, the first will be the flag to set, and the second the value
- in case of bollean flags, only one argument is usfficient, which will toggle the state
- if no argument is given, simply print the state/value of all flags

ASSISTANT: To implement the 'set' subcommand, we need to modify the `cmd/set.go` file.

The 'set' subcommand will have the following behavior:
- If two arguments are supplied, the first will be the flag to set, and the second will be the value to set it to.
- If a boolean flag is supplied with no value, it will toggle the state of the flag.
- If no arguments are given, it will print the state/value of all flags.

Here is the code to implement this:

```go
cmd/set.go
<<<<<<< HEAD
=======
package cmd

import (
	"fmt"
	"strconv"

	"github.com/spf13/cobra"
)

var setCmd = &cobra.Command{
	Use:   "set [flag] [value]",
	Short: "Set the value of a flag",
	Long: `Set the value of a flag. If no value is provided for a boolean flag, it will toggle the state of the flag.
If no arguments are given, it will print the state/value of all flags.`,
	Args: cobra.MaximumNArgs(2),
	Run: func(cmd *cobra.Command, args []string) {
		if len(args) == 0 {
			fmt.Printf("number-lines: %d\n", numLines)
			fmt.Printf("interactive: %t\n", interactive)
			fmt.Printf("start: %s\n", start)
			fmt.Printf("end: %s\n", end)
			fmt.Printf("group-by: %s\n", groupBy)
		} else if len(args) == 1 {
			switch args[0] {
			case "interactive":
				interactive = !interactive
			default:
				fmt.Printf("Unknown flag: %s\n", args[0])
			}
		} else {
			switch args[0] {
			case "number-lines":
				num, err := strconv.Atoi(args[1])
				if err != nil {
					fmt.Printf("Invalid value for number-lines: %s\n", args[1])
				} else {
					numLines = num
				}
			case "start":
				start = args[1]
			case "end":
				end = args[1]
			case "group-by":
				groupBy = args[1]
			default:
				fmt.Printf("Unknown flag: %s\n", args[0])
			}
		}
	},
}

func init() {
	rootCmd.AddCommand(setCmd)
}
>>>>>>> updated
```

This code creates a new 'set' command and adds it to the root command. The 'set' command takes up to two arguments. If no arguments are given, it prints the current state of all flags. If one argument is given, it toggles the state of the 'interactive' flag (if that's the flag provided). If two arguments are given, it sets the value of the specified flag to the provided value.