aider: Updated error message for invalid field in group-by flag.
# Aider chat conversation:

USER: [

ASSISTANT: I'm sorry, but your request is not clear. Could you please provide more details or clarify what you need?

USER: - please output the valid fields as a helpful error message if the user didn't enter a valid entry for group-by

ASSISTANT: To implement this change, we need to modify the `set.go` file. Specifically, we will update the error message in the `group-by` case of the switch statement inside the `Run` function of `setCmd`.

Here is the *edit block* for this change:

```go
cmd/set.go
<<<<<<< HEAD
			case "group-by":
				validFields := []string{"IP", "Timestamp", "StatusCode", "BytesSent", "RequestMethod", "RequestURL", "RequestProtocol", "Referrer", "UserAgent", "Checksum"}
				if contains(validFields, args[1]) {
					GroupBy = args[1]
				} else {
					fmt.Printf("Invalid field for group-by: %s\n", args[1])
				}
=======
			case "group-by":
				validFields := []string{"IP", "Timestamp", "StatusCode", "BytesSent", "RequestMethod", "RequestURL", "RequestProtocol", "Referrer", "UserAgent", "Checksum"}
				if contains(validFields, args[1]) {
					GroupBy = args[1]
				} else {
					fmt.Printf("Invalid field for group-by: %s. Valid fields are: %s\n", args[1], strings.Join(validFields, ", "))
				}
>>>>>>> updated
```

This change will print out the valid fields as part of the error message when the user enters an invalid field for the `group-by` flag.