aider: Refactor input check for "group-by" flag to be case-insensitive and set the variable correctly to the actual field name of the Nginx log.
# Aider chat conversation:

USER: please make the input check for the group-by more user-friendly: it should accept input in any capitalisation, but set the var correctly to the actual field name of the NginxA

ASSISTANT: To make the input check for the "group-by" flag more user-friendly, we need to modify the `setCmd` function in the `cmd/set.go` file.

We will change the `case "group-by"` block to convert the input to uppercase before checking if it's valid. If it is, we will set the `GroupBy` variable to the correct field name in the original capitalization.

Here is the *edit block* for the changes:

```go
cmd/set.go
<<<<<<< HEAD
			case "group-by":
				validFields := []string{"IP", "Timestamp", "StatusCode", "BytesSent", "RequestMethod", "RequestURL", "RequestProtocol", "Referrer", "UserAgent", "Checksum"}
				if contains(validFields, args[1]) {
					GroupBy = args[1]
				} else {
					fmt.Printf("Invalid field for group-by: %s. Valid fields are: %s\n", args[1], strings.Join(validFields, ", "))
				}
=======
			case "group-by":
				validFields := map[string]string{
					"IP": "IP", "TIMESTAMP": "Timestamp", "STATUSCODE": "StatusCode", "BYTESSENT": "BytesSent",
					"REQUESTMETHOD": "RequestMethod", "REQUESTURL": "RequestURL", "REQUESTPROTOCOL": "RequestProtocol",
					"REFERRER": "Referrer", "USERAGENT": "UserAgent", "CHECKSUM": "Checksum",
				}
				upperArg := strings.ToUpper(args[1])
				if field, ok := validFields[upperArg]; ok {
					GroupBy = field
				} else {
					fmt.Printf("Invalid field for group-by: %s. Valid fields are: %s\n", args[1], strings.Join(validFields, ", "))
				}
>>>>>>> updated
```

This change will make the "group-by" flag case-insensitive, while preserving the correct capitalization for the `GroupBy` variable.