aider: Add 'top' subcommand to the `RootCmd` instead of `analyzeCmd`.
# Aider chat conversation:

USER: Ah, sorry, small misunderstanding: the should be a sub command to the `RootCmd`. We don not need an a`analyzeCmd`

ASSISTANT: Understood. We need to add the 'top' subcommand to the `RootCmd` instead of `analyzeCmd`.

Here is the code change:

```go
cmd/analyze.go
<<<<<<< HEAD
func init() {
	analyzeCmd.AddCommand(topCmd)
}
=======
func init() {
	RootCmd.AddCommand(topCmd)
}
>>>>>>> updated
```

This code change will add the 'top' command as a subcommand to the `RootCmd` instead of `analyzeCmd`.