aider: Refactored `GetValidFieldNames` function to dynamically generate the list of valid field names using the `NginxAccessLog` struct.
# Aider chat conversation:

USER: Next, let's refactor the `GetValidFieldNames` function. This should use an actual struct of type `NginxAccessLog` to find the valid names

ASSISTANT: To refactor the `GetValidFieldNames` function, we will need to modify the `cmd/set.go` file.

The `GetValidFieldNames` function currently uses a hardcoded map of valid field names. We will change this to use the `NginxAccessLog` struct from the `cmd/root.go` file to dynamically generate the list of valid field names.

We will use the `reflect` package in Go to get the field names from the `NginxAccessLog` struct.

Here are the steps we will follow:

1. Import the `reflect` package at the top of the `cmd/set.go` file.
2. Modify the `GetValidFieldNames` function to use the `NginxAccessLog` struct to get the field names.

Here are the proposed changes:

```go
cmd/set.go
<<<<<<< HEAD
func GetValidFieldNames() []string {
	validFields := map[string]string{
		"IP": "IP", "TIMESTAMP": "Timestamp", "STATUSCODE": "StatusCode", "BYTESSENT": "BytesSent",
		"REQUESTMETHOD": "RequestMethod", "REQUESTURL": "RequestURL", "REQUESTPROTOCOL": "RequestProtocol",
		"REFERRER": "Referrer", "USERAGENT": "UserAgent", "CHECKSUM": "Checksum",
	}
	keys := make([]string, 0, len(validFields))
	for k := range validFields {
		keys = append(keys, k)
	}
	return keys
}
=======
import "reflect"

func GetValidFieldNames() []string {
	e := reflect.ValueOf(&NginxAccessLog{}).Elem()
	var fieldNames []string
	for i := 0; i < e.NumField(); i++ {
		fieldNames = append(fieldNames, e.Type().Field(i).Name)
	}
	return fieldNames
}
>>>>>>> updated
```

Please note that the `NginxAccessLog` struct is defined in the `cmd/root.go` file. If the `cmd/set.go` file does not have access to this struct, we will need to move the struct to a common location where both files can access it.