aider: Change the default value of `numLines` to 20 in `cmd/root.go`.
# Aider chat conversation:

USER: add `cmd/root.go`

ASSISTANT: Sure, now `cmd/root.go` is in the list of *read-write* files. You can propose changes to this file.

USER: please change the default value of number-lines to 20

ASSISTANT: Sure, we can change the default value of `numLines` to 20 in the `init()` function in the `cmd/root.go` file.

Here is the *edit block* for the change:

```go
cmd/root.go
<<<<<<< HEAD
	RootCmd.PersistentFlags().IntVarP(&numLines, "number-lines", "n", 0, "Number of lines to read from the file or stdin")
=======
	RootCmd.PersistentFlags().IntVarP(&numLines, "number-lines", "n", 20, "Number of lines to read from the file or stdin")
>>>>>>> updated
```